Bug 119844 - [From Symphony]The Spin effect's amount property losts when saved pptx to odp

+++ oox/source/ppt/timenodelistcontext.cxx
@@ -804,7 +804,7 @@ namespace oox { namespace ppt {
                 if(attribs.hasAttribute( XML_by ) )
                 {
                     sal_Int32 nBy = attribs.getInteger( XML_by, 0 );
-                    pNode->setBy( makeAny( nBy ) );
+                    pNode->setBy( makeAny( (double)nBy ) );
                 }
                 if(attribs.hasAttribute( XML_from ) )
                 {
